Aniskiki

+++ src/controller/v1/book-controller.mjs
@@ -1,18 +1,24 @@
 import bookRepository from "../../repositories/book-repository.mjs";
+import categoryRepository from "../../repositories/category-repository.mjs";
 import "dotenv/config";
 
 export async function createBook(req, res) {
     try {
         const { title, authors, publishedDate, description, categories } = req.body;
+        const user = req.user.id;
 
-        const existing = await bookRepository.getbookByTitle(title); //Verificar si ya existe el titulo
+        const existing = await bookRepository.getBookByTitleAndAuthor(title, authors); //Verificar si ya existe el titulo
         if (existing) return res.status(400).json({ message: "Libro ya existente" });
+
+        const category = await categoryRepository.getCategoryById(categories); //Verificar si la categoría existe
+        if (!category) return res.status(404).json({ message: "Categoría no encontrada" });
 
         const book = await bookRepository.create({  //Crear libro
             title, 
             authors, 
             publishedDate, 
             description, 
+            userId: user,
             categories
         });
         res.status(201).json({
@@ -27,11 +33,24 @@ export async function createBook(req, res) {
         res.status(500).json({ error: err.message });
     }
 }
+export async function getListOfBooks(req, res) {
+    try {
+        const books = await bookRepository.getBooksByUser(req.user.id); //depende de si es todos los libros o los del usuario
+        if (!books || books.length === 0) return res.status(200).json({ message: "El usuario no tiene libros" });
+        res.status(200).json(books);
+    } catch (err) {
+        res.status(500).json({ error: err.message });
+    }
+}
 export async function deleteBook(req, res) {
     try {
         const { id } = req.params;
-        const book = await bookRepository.delete(id);
+        const userId = req.user.id;
+        const book = await bookRepository.getBookByTitleAndAuthor(title, authors); 
         if (!book) return res.status(404).json({ message: "Libro no encontrado" });
+        if (book.userId.toString() !== userId) return res.status(403).json({ message: "No autorizado para eliminar este libro" });
+
+        await bookRepository.delete(id);
         res.status(200).json({ message: "Libro eliminado correctamente" });
     } catch (err) {
         res.status(500).json({ error: err.message });
@@ -41,5 +60,3 @@ export async function deleteBook(req, res) {
 /* Editar libro */
 
 /* Detalles de libro */
-
-/* Listar libros */
